Module aws_sdk_opsworkscm::error

Modules

invalid_next_token_exception

See InvalidNextTokenException

invalid_state_exception

See InvalidStateException

limit_exceeded_exception

See LimitExceededException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_not_found_exception

See ResourceNotFoundException

validation_exception

See ValidationException

Structs

AssociateNodeError

CreateBackupError

CreateServerError

DeleteBackupError

DeleteServerError

DescribeAccountAttributesError

DescribeBackupsError

DescribeEventsError

DescribeNodeAssociationStatusError

DescribeServersError

DisassociateNodeError

ExportServerEngineAttributeError

InvalidNextTokenException

This occurs when the provided nextToken is not valid.

InvalidStateException

The resource is in a state that does not allow you to perform a specified action.

LimitExceededException

The limit of servers or backups has been reached.

ListTagsForResourceError

ResourceAlreadyExistsException

The requested resource cannot be created because it already exists.

ResourceNotFoundException

The requested resource does not exist, or access was denied.

RestoreServerError

StartMaintenanceError

TagResourceError

UntagResourceError

UpdateServerEngineAttributesError

UpdateServerError

ValidationException

One or more of the provided request parameters are not valid.

Enums

AssociateNodeErrorKind

CreateBackupErrorKind

CreateServerErrorKind

DeleteBackupErrorKind

DeleteServerErrorKind

DescribeAccountAttributesErrorKind

DescribeBackupsErrorKind

DescribeEventsErrorKind

DescribeNodeAssociationStatusErrorKind

DescribeServersErrorKind

DisassociateNodeErrorKind

ExportServerEngineAttributeErrorKind

ListTagsForResourceErrorKind

RestoreServerErrorKind

StartMaintenanceErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateServerEngineAttributesErrorKind

UpdateServerErrorKind