Module aws_sdk_opsworkscm::model

Modules

account_attribute

See AccountAttribute

backup

See Backup

engine_attribute

See EngineAttribute

server

See Server

server_event

See ServerEvent

tag

See Tag

Structs

AccountAttribute

Stores account attributes.

Backup

Describes a single backup.

EngineAttribute

A name and value pair that is specific to the engine of the server.

Server

Describes a configuration management server.

ServerEvent

An event that is related to the server, such as the start of maintenance or backup.

Tag

A map that contains tag keys and tag values to attach to an AWS OpsWorks for Chef Automate

Enums

BackupStatus

BackupType

MaintenanceStatus

NodeAssociationStatus

The status of the association or disassociation request.

ServerStatus