Module aws_sdk_opsworkscm::operation

Structs

AssociateNode

Associates a new node with the server. For more information about how to disassociate a node, see DisassociateNode.

CreateBackup

CreateServer

DeleteBackup

DeleteServer

DescribeAccountAttributes

DescribeBackups

DescribeEvents

DescribeNodeAssociationStatus

DescribeServers

DisassociateNode

ExportServerEngineAttribute

ListTagsForResource

Returns a list of tags that are applied to the specified AWS OpsWorks for Chef Automate or

RestoreServer

StartMaintenance

TagResource

Applies tags to an AWS OpsWorks for Chef Automate or AWS OpsWorks for Puppet Enterprise server, or to server backups.

UntagResource

Removes specified tags from an AWS OpsWorks-CM server or backup.

UpdateServer

UpdateServerEngineAttributes